refactor: use global-layout
