+++ lib/layouts/GlobalLayout.vue
@@ -1,0 +1,27 @@
+<template>
+  <div id="app">
+    <TheHeader />
+
+    <TheMain />
+
+    <TheFooter />
+  </div>
+</template>
+
+<script>
+import TheHeader from '../components/TheHeader.vue'
+import TheMain from '../components/TheMain.vue'
+import TheFooter from '../components/TheFooter.vue'
+
+export default {
+  name: 'ThemeMeteorlxy',
+
+  components: {
+    TheHeader,
+    TheMain,
+    TheFooter,
+  },
+}
+</script>
+
+<style lang="stylus" src="../styles/index.styl" />
