+++ lib/layouts/Home.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script>
-import PostsList from '../PostsList.vue'
+import PostsList from '../components/PostsList.vue'
 
 export default {
   name: 'Home',
